+++ pytsite/geo_ip/eh.py
@@ -1,7 +1,7 @@
 """PytSite GeoIP Event Handlers.
 """
 from datetime import datetime as _datetime, timedelta as _timedelta
-from pytsite import odm as _odm, console as _console, db as _db
+from pytsite import odm as _odm, db as _db
 
 __author__ = 'Alexander Shepetko'
 __email__ = 'a@shepetko.com'
@@ -14,7 +14,6 @@ def pytsite_update(version: str):
     if version in ('0.59.0', '0.63.0'):
         # Re-create 'geo_ips' collection due to its structure changes
         _db.get_collection('geo_ips').drop()
-        _console.print_info("Collection 'geo_ips' will be re-created.")
 
 
 def pytsite_cron_weekly():