+++ pytsite/wallet/res/assets/js/widget-money-input.js
@@ -1,4 +1,7 @@
 $(window).on('pytsite.widget.init:pytsite.wallet._widget.MoneyInput', function (e, widget) {
-    // Redirect event to Decimal widget
-    $(window).trigger('pytsite.widget.init:pytsite.widget._input.Decimal', [widget]);
+    widget.em.find('input[type=text],input[type=tel],input[type=number]').inputmask('decimal', {
+        allowMinus: false
+    });
+
+    $(window).trigger('pytsite.widget.init:pytsite.widget._input.Text', [widget]);
 });